requirements.txt
anthropic
xmltodict
argparse
python-dotenv
pyyaml 